Additional range parameter for discerete coordinate spec
This is useful, since not all the types that impls the discerete trait
is discerete by natural which means we might need some additional
information to make it discerete.

For example, a linspace coord, without knowing the information about the
start point and range, we cannot determine which is the next value.

In the future we may want to use a slice as a discerete coord, in this
case the reference to the slice must be used to generate the next value,
etc...