Allow passing colors by value and reference
This fixes a bit of a usability issue with the API of plotters where
certain functions only take references to colors, even if an owned value
is easily available. Especially if the same builder pattern takes
multiple colors, it's extremely annoying if the color cannot be
initialized once and then copied around.