Fix bug #347
This bug was caused by the default rasterizer plolygonize a path
line with stroke width later than 1. When angle is too small, the
polygon vertex might be very far away from the center line.

In SVG rasterizer, this vertex will be dropped - but our rasterizer lack
of this type of check.

This change implemented the check, when distance between the line vertex
the coresponding plolygon vertex on the outter side is larger than 2
times of the stroke width - Then we cap the polygon.