Error propagation is silenced
```
warning: unused `Result` that must be used
   --> plotters/src/style/font/ttf.rs:135:9
    |
135 | /         data.clone().map(|handle| {
136 | |             handle
137 | |                 .load()
138 | |                 .map(FontExt::new)
139 | |                 .map_err(|e| FontError::FontLoadError(Arc::new(e)))
140 | |         })?;
    | |___________^
    |
    = note: this `Result` may be an `Err` variant, which should be handled
    = note: `#[warn(unused_must_use)]` on by default
help: use `let _ = ...` to ignore the resulting value
    |
135 |         let _ = data.clone().map(|handle| {
```

When I remove the ? and look at the type of the droppped variable
I see that is it Result<Result<FontExt, FontError>, FontError>

Rust is very unintuative on this matter. The solution is to double
unwrap.

There are many ways to do this .. I went with a soltions I think is readable.