style/font: use web font style
The second String is not used anywhere so the Rust compiler
will complain like below.

The font-style is actually useful so let's set it.

warning: field `1` is never read
  -->
  /home/adi/workspace/web/deps/plotters/plotters/src/style/font/web.rs:21:37
     |
  21 | pub struct FontDataInternal(String, String);
     |            ----------------         ^^^^^^
     |            field in this struct
     |
     = note: `FontDataInternal` has a derived impl for the trait `Clone`, but this is intentionally ignored during dead code analysis
     = note: `#[warn(dead_code)]` on by default
help: consider changing the field to be of unit type to suppress this warning while preserving the field numbering, or remove the field
     |
  21 | pub struct FontDataInternal(String, ());
     |                                     ~~

warning: `plotters` (lib) generated 1 warning